requirements.txt
# Core dependencies
tqdm>=4.65.0
# Utilities

ipykernel>=6.0.0
jupyter>=1.0.0
# Jupyter

pyyaml>=6.0
# Configuration

seaborn>=0.12.0
matplotlib>=3.7.0
# Visualization

spacy>=3.5.0
nltk>=3.8
# Data processing

transformers>=4.30.0
torch>=2.0.0
# Deep Learning

scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

